requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
google-generativeai
asyncpg
